
\name{BACCO-package}
\Rdversion{1.1}
\alias{BACCO-package}
\alias{BACCO}
\docType{package}
\title{
Bayesian Analysis of Computer Code Output (BACCO)
}
\description{
The BACCO bundle of packages is replaced by the BACCO package, which
provides a vignette that illustrates the constituent packages in use.
}
\details{

This file contains no code: the bundle formerly known as BACCO is
now a \dQuote{virtual} package which requires the constituent packages:
\CRANpkg{emulator}, \CRANpkg{calibrator}, and \CRANpkg{approximator}.
Package \CRANpkg{multivator} is a standalone package.

}

\author{
\packageAuthor{BACCO}

Maintainer: \packageMaintainer{BACCO}
}

\references{
  \itemize{
    \item
    J. Oakley 1999. \emph{Bayesian uncertainty analysis for complex
      computer codes}, PhD thesis, University of Sheffield
    \item
    J. Oakley and A. O'Hagan, 2002. \emph{Bayesian Inference for the
      Uncertainty Distribution of Computer Model Outputs}, Biometrika
    89(4), pp769-784
    \item
    R. K. S. Hankin 2005. \emph{Introducing BACCO, an R bundle for
      Bayesian analysis of computer code output}, Journal of Statistical
    Software, 14(16)
  }
}
\keyword{ package }

